fix: some tests are not pass
