+++ packages/base-docs/src/Controller/inline-format.controller.ts
@@ -1,21 +1,14 @@
-import { ITextSelectionRangeWithStyle, ITextSelectionRenderManager } from '@univerjs/base-render';
+import { ITextSelectionRenderManager } from '@univerjs/base-render';
 import {
-    BooleanNumber,
     Disposable,
-    getTextIndexByCursor,
     ICommandInfo,
     ICommandService,
-    IDocumentBody,
-    IStyleBase,
-    ITextDecoration,
-    ITextRun,
     IUniverInstanceService,
     LifecycleStages,
     OnLifecycle,
 } from '@univerjs/core';
 import { Inject } from '@wendellhu/redi';
 
-import MemoryCursor from '../Basics/memoryCursor';
 import {
     SetInlineFormatBoldCommand,
     SetInlineFormatCommand,
@@ -26,7 +19,6 @@ import {
     SetInlineFormatTextColorCommand,
     SetInlineFormatUnderlineCommand,
 } from '../commands/commands/inline-format.command';
-import { IRichTextEditingMutationParams, RichTextEditingMutation } from '../commands/mutations/core-editing.mutation';
 import { TextSelectionManagerService } from '../services/text-selection-manager.service';
 
 /**
@@ -64,124 +56,22 @@ export class InlineFormatController extends Disposable {
                     return;
                 }
 
-                this.handleInlineFormat(command as ICommandInfo<{ value: string }>);
+                this.handleInlineFormat(command);
             })
         );
     }
 
-    private handleInlineFormat(
-        command: ICommandInfo<{
-            value: string;
-        }>
-    ) {
-        console.log(command);
+    private handleInlineFormat(command: ICommandInfo) {
         const segmentId = this._textSelectionRenderManager.getActiveRange()?.segmentId;
-        const selections = this._textSelectionManagerService.getSelections();
 
-        if (segmentId == null || !Array.isArray(selections) || selections.length === 0) {
+        if (segmentId == null) {
             return;
         }
 
-        const docsModel = this._currentUniverService.getCurrentUniverDocInstance();
-        const unitId = docsModel.getUnitId();
-
-        let formatValue;
-
-        const COMMAND_ID_TO_FORMAT_KEY_MAP: Record<string, keyof IStyleBase> = {
-            [SetInlineFormatBoldCommand.id]: 'bl',
-            [SetInlineFormatItalicCommand.id]: 'it',
-            [SetInlineFormatUnderlineCommand.id]: 'ul',
-            [SetInlineFormatStrikethroughCommand.id]: 'st',
-            [SetInlineFormatFontSizeCommand.id]: 'fs',
-            [SetInlineFormatFontFamilyCommand.id]: 'ff',
-            [SetInlineFormatTextColorCommand.id]: 'cl',
-        };
-
-        switch (command.id) {
-            case SetInlineFormatBoldCommand.id: // fallthrough
-            case SetInlineFormatItalicCommand.id: // fallthrough
-            case SetInlineFormatUnderlineCommand.id: // fallthrough
-            case SetInlineFormatStrikethroughCommand.id: {
-                formatValue = getReverseFormatValueInSelection(
-                    docsModel.body!.textRuns!,
-                    COMMAND_ID_TO_FORMAT_KEY_MAP[command.id],
-                    selections
-                );
-
-                break;
-            }
-
-            case SetInlineFormatFontSizeCommand.id:
-            case SetInlineFormatFontFamilyCommand.id: {
-                formatValue = command.params?.value;
-                break;
-            }
-
-            case SetInlineFormatTextColorCommand.id: {
-                formatValue = {
-                    rgb: command.params?.value,
-                };
-                break;
-            }
-
-            default: {
-                throw new Error(`Unknown command: ${command.id} in handleInlineFormat`);
-            }
-        }
-
-        const doMutation: ICommandInfo<IRichTextEditingMutationParams> = {
-            id: RichTextEditingMutation.id,
-            params: {
-                unitId,
-                mutations: [],
-            },
-        };
-
-        const memoryCursor = new MemoryCursor();
-
-        memoryCursor.reset();
-
-        for (const selection of selections) {
-            const { cursorStart, cursorEnd, isStartBack, isEndBack } = selection;
-            const textStart = getTextIndexByCursor(cursorStart, isStartBack);
-            const textEnd = getTextIndexByCursor(cursorEnd, isEndBack);
-
-            const body: IDocumentBody = {
-                dataStream: '',
-                textRuns: [
-                    {
-                        st: 0,
-                        ed: textEnd - textStart,
-                        ts: {
-                            [COMMAND_ID_TO_FORMAT_KEY_MAP[command.id]]: formatValue,
-                        },
-                    },
-                ],
-            };
-
-            const len = textStart + 1 - memoryCursor.cursor;
-            if (len !== 0) {
-                doMutation.params!.mutations.push({
-                    t: 'r',
-                    len,
-                    segmentId,
-                });
-            }
-
-            doMutation.params!.mutations.push({
-                t: 'r',
-                body,
-                len: textEnd - textStart,
-                segmentId,
-            });
-
-            memoryCursor.reset();
-            memoryCursor.moveCursor(textEnd + 1);
-        }
-
         this._commandService.executeCommand(SetInlineFormatCommand.id, {
-            unitId,
-            doMutation,
+            segmentId,
+            preCommandId: command.id,
+            ...(command.params ?? {}),
         });
 
         const REFRESH_SELECTION_COMMAND_LIST = [
@@ -190,68 +80,11 @@ export class InlineFormatController extends Disposable {
             SetInlineFormatFontFamilyCommand.id,
         ];
 
+        /**
+         * refresh selection.
+         */
         if (REFRESH_SELECTION_COMMAND_LIST.includes(command.id)) {
             this._textSelectionManagerService.refreshSelection();
         }
     }
-}
-
-function isTextDecoration(value: unknown | ITextDecoration): value is ITextDecoration {
-    return value !== null && typeof value === 'object';
-}
-
-/**
- * When clicking on a Bold menu item, you should un-bold if there is bold in the selections,
- * or bold if there is no bold text. This method is used to get the style value calculated
- * from textRuns in the selection
- */
-function getReverseFormatValueInSelection(
-    textRuns: ITextRun[],
-    key: keyof IStyleBase,
-    selections: ITextSelectionRangeWithStyle[]
-): BooleanNumber | ITextDecoration {
-    let ti = 0;
-    let si = 0;
-
-    while (ti !== textRuns.length && si !== selections.length) {
-        const { cursorStart, cursorEnd, isStartBack, isEndBack } = selections[si];
-
-        const textStart = getTextIndexByCursor(cursorStart, isStartBack) + 1;
-        const textEnd = getTextIndexByCursor(cursorEnd, isEndBack) + 1;
-
-        // TODO: @jocs handle sid in textRun
-        const { st, ed, ts } = textRuns[ti];
-
-        if (textEnd <= st) {
-            si++;
-        } else if (ed <= textStart) {
-            ti++;
-        } else {
-            if (ts?.[key] == null) {
-                return /bl|it/.test(key)
-                    ? BooleanNumber.TRUE
-                    : {
-                          s: BooleanNumber.TRUE,
-                      };
-            }
-
-            if (ts[key] === BooleanNumber.FALSE) {
-                return BooleanNumber.TRUE;
-            }
-
-            if (isTextDecoration(ts[key]) && (ts[key] as ITextDecoration).s === BooleanNumber.FALSE) {
-                return {
-                    s: BooleanNumber.TRUE,
-                };
-            }
-
-            ti++;
-        }
-    }
-
-    return /bl|it/.test(key)
-        ? BooleanNumber.FALSE
-        : {
-              s: BooleanNumber.FALSE,
-          };
 }
