+++ packages/base-docs/src/commands/commands/__tests__/inline-format.command.spec.ts
@@ -12,8 +12,10 @@ import {
 import { Injector } from '@wendellhu/redi';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
+import { NORMAL_TEXT_SELECTION_PLUGIN_NAME } from '../../../Basics/docs-view-key';
+import { TextSelectionManagerService } from '../../../services/text-selection-manager.service';
 import { RichTextEditingMutation } from '../../mutations/core-editing.mutation';
-import { SetInlineFormatCommand } from '../inline-format.command';
+import { SetInlineFormatBoldCommand, SetInlineFormatCommand } from '../inline-format.command';
 import { createCommandTestBed } from './create-command-test-bed';
 
 describe('Test inline format commands', () => {
@@ -46,27 +48,26 @@ describe('Test inline format commands', () => {
         get = testBed.get;
 
         commandService = get(ICommandService);
-        // commandService.registerCommand(SetInlineFormatBoldCommand);
         commandService.registerCommand(SetInlineFormatCommand);
         commandService.registerCommand(RichTextEditingMutation as unknown as ICommand);
 
-        // const selectionManager = get(TextSelectionManagerService);
+        const selectionManager = get(TextSelectionManagerService);
 
-        // selectionManager.setCurrentSelection({
-        //     pluginName: NORMAL_TEXT_SELECTION_PLUGIN_NAME,
-        //     unitId: 'test-doc',
-        // });
+        selectionManager.setCurrentSelection({
+            pluginName: NORMAL_TEXT_SELECTION_PLUGIN_NAME,
+            unitId: 'test-doc',
+        });
 
-        // selectionManager.add([
-        //     {
-        //         cursorStart: 0,
-        //         isStartBack: true,
-        //         cursorEnd: 5,
-        //         isEndBack: false,
-        //         isCollapse: false,
-        //         segmentId: '',
-        //     },
-        // ]);
+        selectionManager.add([
+            {
+                cursorStart: 0,
+                isStartBack: true,
+                cursorEnd: 5,
+                isEndBack: false,
+                isCollapse: false,
+                segmentId: '',
+            },
+        ]);
     });
 
     afterEach(() => univer.dispose());
@@ -75,36 +76,12 @@ describe('Test inline format commands', () => {
         it('Should change text in range(0, 5) to bold', async () => {
             expect(getFormatValueAt('bl', 1)).toBe(BooleanNumber.FALSE);
 
-            const doMutation = {
-                unitId: 'test-doc',
-                doMutation: {
-                    id: 'doc.mutation.rich-text-editing',
-                    params: {
-                        unitId: 'test-doc',
-                        mutations: [
-                            {
-                                t: 'r',
-                                body: {
-                                    dataStream: '',
-                                    textRuns: [
-                                        {
-                                            st: 0,
-                                            ed: 5,
-                                            ts: {
-                                                bl: BooleanNumber.TRUE,
-                                            },
-                                        },
-                                    ],
-                                },
-                                len: 5,
-                                segmentId: '',
-                            },
-                        ],
-                    },
-                },
+            const commandParams = {
+                segmentId: '',
+                preCommandId: SetInlineFormatBoldCommand.id,
             };
 
-            await commandService.executeCommand(SetInlineFormatCommand.id, doMutation);
+            await commandService.executeCommand(SetInlineFormatCommand.id, commandParams);
 
             expect(getFormatValueAt('bl', 1)).toBe(BooleanNumber.TRUE);
